Fix 2

+++ tasks/seq/borisov_s_strassen_seq/func_tests/main.cpp
@@ -445,9 +445,9 @@ TEST(borisov_s_strassen_seq, Square240x240_Random) {
 }
 
 TEST(borisov_s_strassen_seq, ValidCase) {
-  std::vector<int> input_data = {2, 3, 3, 2};
-  input_data.insert(input_data.end(), {1, 2, 3, 4, 5, 6});
-  input_data.insert(input_data.end(), {7, 8, 9, 10, 11, 12});
+  std::vector<double> input_data = {2.0, 3.0, 3.0, 2.0};
+  input_data.insert(input_data.end(), {1.0, 2.0, 3.0, 4.0, 5.0, 6.0});
+  input_data.insert(input_data.end(), {7.0, 8.0, 9.0, 10.0, 11.0, 12.0});
 
   auto task_data = std::make_shared<ppc::core::TaskData>();
   task_data->inputs.push_back(reinterpret_cast<uint8_t*>(input_data.data()));
